refactor(myweb): extract the force layout into useWebGraphSimulation
Move the d3-force simulation — building and rebuilding it over the filtered
subgraph, painting normalized positions into `nodes`, fitting the viewport
as it settles, and node dragging — plus all its refs and FIT_VIEW_PADDING
out of web-graph.tsx into a hook. WebGraph passes the filtered subgraph in
and renders the `nodes` it returns, wiring onInit → registerFlow, onMove →
markUserMoved, and the mode-toggle refit → fitView. Trims ~320 lines from
the component, quarantining its densest, most imperative code. Pure move;
the layout math stays unit-tested in web-graph-layout.ts.
